Use new deployment.can for the <DeploymentMenu />
