+++ src/components/DeploymentMenu.vue
@@ -1,16 +1,16 @@
 <template>
   <p-icon-button-menu v-bind="$attrs">
-    <DeploymentQuickRunOverflowMenuItem v-if="can.run.deployment && showAll" :deployment="deployment" :open-modal="openParametersModal" />
+    <DeploymentQuickRunOverflowMenuItem v-if="deployment.can.run && showAll" :deployment="deployment" :open-modal="openParametersModal" />
 
-    <DeploymentCustomRunOverflowMenuItem v-if="can.run.deployment && showAll" :deployment-id="deployment.id" />
+    <DeploymentCustomRunOverflowMenuItem v-if="deployment.can.run && showAll" :deployment-id="deployment.id" />
 
     <copy-overflow-menu-item label="Copy ID" :item="deployment.id" />
 
-    <router-link v-if="!deployment.deprecated && can.update.deployment" :to="routes.deploymentEdit(deployment.id)">
+    <router-link v-if="!deployment.deprecated && deployment.can.manage" :to="routes.deploymentEdit(deployment.id)">
       <p-overflow-menu-item label="Edit" />
     </router-link>
 
-    <p-overflow-menu-item v-if="can.delete.deployment" label="Delete" @click="openConfirmDeleteModal" />
+    <p-overflow-menu-item v-if="deployment.can.manage" label="Delete" @click="openConfirmDeleteModal" />
 
     <slot v-bind="{ deployment }" />
   </p-icon-button-menu>
@@ -36,19 +36,17 @@
 <script lang="ts" setup>
   import { DeploymentQuickRunOverflowMenuItem, DeploymentCustomRunOverflowMenuItem, ConfirmDeleteModal, CopyOverflowMenuItem, QuickRunParametersModal } from '@/components'
   import { useWorkspaceApi, useWorkspaceRoutes, useCan, useShowModal } from '@/compositions'
-  import { Deployment } from '@/models'
+  import { DeploymentWithObjectLevelScopes } from '@/models/DeploymentWithObjectLevelScopes'
   import { deleteItem } from '@/utilities'
 
   defineProps<{
-    deployment: Deployment,
+    deployment: DeploymentWithObjectLevelScopes,
     showAll?: boolean,
   }>()
 
   const emits = defineEmits<{
     (event: 'delete', value: string): void,
   }>()
-
-  const can = useCan()
 
   const { showModal: showConfirmDeleteModal, open: openConfirmDeleteModal, close: closeConfirmDeleteModal } = useShowModal()
   const { showModal: showParametersModal, open: openParametersModal } = useShowModal()
